fix: resolve circular dependency
There is
