+++ src/app/app.html
@@ -156,7 +156,7 @@
 
     <mat-menu #menuApps="matMenu">
       @if (app.authenticated()) {
-      <button mat-menu-item (click)="layout.createNote()">
+      <button mat-menu-item (click)="eventService.createNote()">
         <mat-icon>create</mat-icon>
         <span>Create Note</span>
       </button>
@@ -293,7 +293,7 @@
       <!-- <mat-toolbar>
         <span>Profile</span>
         <span class="toolbar-spacer"></span>
-        
+
       </mat-toolbar> -->
       <div class="profile-content">
         <button
@@ -479,7 +479,7 @@
   <button mat-mini-fab color="primary" class="desktop-create-fab no-print" aria-label="Create new content"
     (click)="openCreateOptions()">
     <mat-icon>add</mat-icon>
-  </button> 
+  </button>
   } -->
 
   <!-- Mobile navigation footer - now outside of sidenav container -->
